move cropper logic into component
